% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdg.R
\name{cdg_choose_millesime}
\alias{cdg_choose_millesime}
\title{Choose available years (millesimes)}
\usage{
cdg_choose_millesime(site, millesime = NULL, ...)
}
\arguments{
\item{site}{\code{character} string specifying the data site. Passed to \code{\link{cdg_detect_millesimes}}.}

\item{millesime}{\code{character} specifying the desired millesime:
\itemize{
\item \code{NULL} → returns \code{"latest"}.
\item \code{"?"} → opens an interactive menu to choose from available millesimes.
\item otherwise, checks if the provided millesime exists.
}}

\item{...}{Additional arguments passed to \code{\link{cdg_detect_millesimes}}.}
}
\value{
A \code{character} string representing the chosen millesime, or \code{"latest"} if not found or canceled.
}
\description{
Allows the user to select a specific "millesime" (data year version) for a given site.
If no millesime is specified, the function defaults to \code{"latest"}.
}
\details{
\itemize{
\item If \code{millesime} is \code{"?"}, the user is prompted with an interactive selection menu.
\item If the chosen millesime is not available, a warning is issued and \code{"latest"} is returned.
}
}
\examples{
\dontrun{
cdg_choose_millesime("pci")
cdg_choose_millesime("pci", "?")
cdg_choose_millesime("pci", "2023")
}

}
\seealso{
\code{\link{cdg_detect_millesimes}}
}
