% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdg.R
\name{cdg_detect_millesimes}
\alias{cdg_detect_millesimes}
\title{Detect available years (millesimes)}
\usage{
cdg_detect_millesimes(site, allowed_sites = c("pci", "etalab"))
}
\arguments{
\item{site}{\code{character}.
String indicating the data site to query. Must be one of \code{"pci"} or \code{"etalab"}.}

\item{allowed_sites}{\code{character}.
Vector of allowed site names. Defaults to \code{c("pci", "etalab")}.}
}
\value{
A \code{character} vector of unique year identifiers (millesimes) found on the site.
}
\description{
Retrieves and returns the list of available "millesimes" (year directories) from a specified data site.
}
\details{
This function uses \code{cdg_get_path} to get the base URL for the site,
then calls \code{cdg_detect_links} to list all hyperlinks on the site,
filters those that represent directories (ending with a slash),
and returns the unique directory names without the trailing slash.
}
\examples{
\dontrun{
years <- cdg_detect_millesimes("pci")
print(years)
}

}
\seealso{
\code{\link[=cdg_detect_links(), cdg_get_path]{cdg_detect_links(), cdg_get_path()}}
}
