% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cfg_get_config_path}
\alias{cfg_get_config_path}
\title{Get the path to a configuration file in the package}
\usage{
cfg_get_config_path(cfg_name = NULL, must_exist = TRUE)
}
\arguments{
\item{cfg_name}{\code{character} or \code{NULL}.
Name of the configuration file. If \code{NULL}, the directory path is returned.}

\item{must_exist}{\code{logical}. Default is \code{TRUE}.
If \code{TRUE}, throws an error if the configuration file does not exist.}
}
\value{
\code{character}. Full path to the configuration file or the configuration directory.
}
\description{
Returns the full path to a configuration file located in the \verb{inst/config/}
directory of the \code{rcadastre} package. If no file name is provided, returns
the path to the configuration directory.
}
