% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cfg_get_prefix_extent}
\alias{cfg_get_prefix_extent}
\title{Récupère le préfixe et l'extension selon le site, le format et l'échelle}
\usage{
cfg_get_prefix_extent(
  site,
  format = "edigeo",
  scale,
  data = NULL,
  config = cfg_load("cdg_structure.yaml")
)
}
\arguments{
\item{site}{\code{character}.
Nom du site, par ex. "pci" ou "etalab".}

\item{format}{\code{character}.
Nom du format, par ex. "dxf", "edigeo", "shp", "geojson".}

\item{scale}{\code{character}.
Échelle ou subdivision, par ex. "departements", "feuilles", "communes".}

\item{data}{\code{character}, \code{NULL} par défaut.
Le type de donnée, par ex. "raw", "proc"}

\item{config}{\code{list}.
Configuration YAML déjà chargée (par défaut via cfg_load()).}
}
\value{
Une liste contenant \code{prefix} et \code{extent}, ou d'autres champs s'ils sont définis.
}
\description{
Récupère le préfixe et l'extension selon le site, le format et l'échelle
}
\examples{
\dontrun{
cfg <- cfg_load("cdg_structure.yaml")
cfg_get_prefix_extent("pci", "dxf", "departements", config = cfg)
cfg_get_prefix_extent("etalab", "geojson", "communes", config = cfg)
}
}
