% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cfg_load}
\alias{cfg_load}
\title{Load a YAML configuration file from the package}
\usage{
cfg_load(cfg_name)
}
\arguments{
\item{cfg_name}{\code{character}. Name of the YAML configuration file to load.}
}
\value{
\code{list}. Contents of the YAML configuration file.
}
\description{
Reads a YAML configuration file located in the \verb{inst/config/} directory of
the \code{rcadastre} package and returns it as a list.
}
\seealso{
\code{\link[=cfg_get_config_path]{cfg_get_config_path()}}
}
