% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cog_2025.R
\docType{data}
\name{departement_2025}
\alias{departement_2025}
\title{French Departments Data}
\format{
A data frame with 6 columns:
\describe{
\item{REG}{(integer) Region code, 2 digits.}
\item{DEP}{(character) Department code, 3 characters.}
\item{CHEFLIEU}{(character) Department capital code, 5 digits.}
\item{TNCC}{(integer) Type of clear name, 1 digit.}
\item{NCC}{(character) Clear name (uppercase), up to 200 characters.}
\item{NCCENR}{(character) Clear name with rich typography, up to 200 characters.}
\item{LIBELLE}{(character) Clear name with rich typography and article, up to 200 characters.}
}
}
\source{
\url{https://www.insee.fr/fr/information/8377162}
}
\usage{
departement_2025
}
\description{
A dataset containing information about French departments, including their codes and names.
}
\keyword{datasets}
