% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_download.R
\name{.extract_gz}
\alias{.extract_gz}
\title{Extract a .gz file}
\usage{
.extract_gz(src, dest)
}
\arguments{
\item{src}{Character, path to the .gz file}

\item{dest}{Character, path to write the decompressed file}
}
\value{
Path to the decompressed file
}
\description{
Extract a .gz file
}
\keyword{internal}
