% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etalab.R
\name{.get_etalab_arg_pairs}
\alias{.get_etalab_arg_pairs}
\title{Generate Commune-Layer Pairs for Etalab Data}
\usage{
.get_etalab_arg_pairs(commune, data)
}
\arguments{
\item{commune}{\code{character} vector. The INSEE code(s) of the commune(s).}

\item{data}{\code{character} vector or \code{list}.
If \code{character}, a vector of layers will be paired with all communes (Cartesian product).
If \code{list}, each element corresponds to a vector of layers for the matching commune.}
}
\value{
A \code{data.frame} with columns:
\itemize{
\item \code{commune}: commune code
\item \code{layer}: corresponding layer
}
}
\description{
Generates a data.frame of commune-layer pairs to facilitate Etalab data queries.
}
\details{
\itemize{
\item If \code{data} is a character vector, all layers are combined with all communes.
\item If \code{data} is a list, each element is paired with the corresponding commune.
\item The function ensures that a list of layers matches the length of \code{commune}.
}
}
\examples{
\dontrun{
# Cartesian product example
.get_etalab_arg_pairs(c("72187", "72188"), c("parcelles", "lieux_dits"))

# Pairwise example
.get_etalab_arg_pairs(c("72187", "72188"), list(
  c("parcelles", "lieux_dits"),
  c("parcelles")
))
}

}
\keyword{internal}
