% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etalab.R
\name{.get_etalab_urls}
\alias{.get_etalab_urls}
\title{Generate URLs for Etalab Cadastre Data}
\usage{
.get_etalab_urls(commune, data, millesime = "latest")
}
\arguments{
\item{commune}{\code{character} or \code{numeric} vector. The INSEE code(s) of the commune(s).}

\item{data}{\code{character} vector or \code{list}.
If \code{character}, a vector of layers will be paired with all communes (Cartesian product).
If \code{list}, each element corresponds to a vector of layers for the matching commune.}

\item{millesime}{\code{character}. The version or millesime of the dataset.
Must be on of \code{get_data_millesimes("etalab")}. Default is \code{"latest"}.}
}
\value{
A \code{character} vector containing unique URLs to the requested Etalab data layers.
}
\description{
This internal function constructs URLs for Etalab cadastre data
given a set of commune codes and requested layers.
}
\details{
\itemize{
\item Validates that the requested layers exist.
\item Generates commune-layer pairs (cartesian or pairwise).
\item Determines layer type (\code{raw} or \code{proc}) and constructs the corresponding URLs.
\item Only URLs matching the expected pattern are returned.
}
}
\examples{
\dontrun{
# Single commune, multiple layers
.get_etalab_urls("72187", c("parcelles", "lieux_dits"))

# Multiple communes, pairwise layers
.get_etalab_urls(c("72187", "72181"), list(c("parcelles", "lieux_dits"),c("commune")))
}

}
\seealso{
\code{\link[=get_etalab_layernames]{get_etalab_layernames()}}, \code{\link[=get_data_millesimes]{get_data_millesimes()}}
}
\keyword{internal}
