% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_download.R
\name{.get_extension}
\alias{.get_extension}
\title{Get file extension}
\usage{
.get_extension(filename)
}
\arguments{
\item{filename}{Character, file name or path}
}
\value{
File extension in lower case
}
\description{
Get file extension
}
\keyword{internal}
