% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pci.R
\name{.get_pci_urls}
\alias{.get_pci_urls}
\title{Generate PCI sheet URLs from Commune or Sheet codes}
\usage{
.get_pci_urls(x, millesime = "latest", format = "edigeo")
}
\arguments{
\item{x}{\code{character}. Vector of codes to retrieve URLs for.
Each element must be either a 5-character commune code or a 12-character sheet code.}

\item{millesime}{\code{character}. The version or millesime of the dataset.
Must be on of \code{get_data_millesimes("pci")}. Default is \code{"latest"}.}

\item{format}{\code{character}. The format of the data.
Must be \code{"edigeo"} or \code{"dxf"}. Default is \code{"edigeo"}.}
}
\value{
A character vector of full URLs pointing to the requested PCI sheets.
}
\description{
This internal function generates PCI (Plan Cadastral Informatise) URLs
given a vector of commune codes (5-character) or individual sheet codes
(12-character).
}
\details{
\itemize{
\item If an element of \code{x} is a 5-character code, all sheets for that commune are returned.
\item If an element of \code{x} is a 12-character code, only the URL for that specific sheet is returned.
\item Invalid codes will throw an error.
}
}
\examples{
\dontrun{
# Generate URLs for a single commune
.get_pci_urls("72187")

# Generate URL for a specific sheet
.get_pci_urls("72181000AB01")

# Multiple codes at once
.get_pci_urls(c("72187", "72181000AB01"))
}

}
\seealso{
\code{\link[=get_data_millesimes]{get_data_millesimes()}}
}
\keyword{internal}
