% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idu.R
\name{.idu_assert}
\alias{.idu_assert}
\title{Assert that IDU codes are valid}
\usage{
.idu_assert(idu)
}
\arguments{
\item{idu}{Character vector of IDU codes.}
}
\value{
Invisibly returns TRUE if all IDUs are valid.
}
\description{
Throws an error if any element of \code{idu} is invalid.
}
\keyword{internal}
