% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idu.R
\name{.idu_get_contenance}
\alias{.idu_get_contenance}
\title{Retrieve contenance values for given IDUs}
\usage{
.idu_get_contenance(idu)
}
\arguments{
\item{idu}{A \code{character} vector of valid IDU codes.}
}
\value{
A \code{data.frame} containing the split IDU parts and an additional
\code{contenance} column with the matched surfaces.
}
\description{
This function takes one or more valid IDU codes, retrieves parcel
data from Etalab, and returns a data.frame linking each IDU to its \code{contenance}
(surface area in square meters).
}
\examples{
\dontrun{
# Example IDUs
my_idus <- c("12345ABCDE6789", "54321ZZZZZ1234")

# Get parcel contenance
contenance_df <- idu_get_contenance(my_idus)
print(contenance_df)
}

}
\keyword{internal}
