% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idu.R
\name{.idu_get_name}
\alias{.idu_get_name}
\title{Get region/department/commune names from IDU codes}
\usage{
.idu_get_name(idu, loc = c("both", "reg", "dep", "com"), cog_field = "NCC")
}
\arguments{
\item{idu}{A \code{character} vector of valid IDU codes.}

\item{loc}{A \code{character} string specifying which location levels to include.
One of \code{"both"}, \code{"reg"}, \code{"dep"}, or \code{"com"}.}

\item{cog_field}{\code{character}. Défaut is \code{"NCC"}.
The name of the field in the reference datasets to use.}
}
\value{
A \code{data.frame} with the IDU split into its components and
the requested location names.
}
\description{
This function takes one or more IDU codes, validates them, splits them into
their components, and merges them with reference datasets to retrieve the
corresponding region, department, and/or commune names.
}
\examples{
# idu_get_name(c("12345ABCDE6789", "54321ZZZZZ1234"), loc = "both")

}
\keyword{internal}
