% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_download.R
\name{.log_msg}
\alias{.log_msg}
\title{Print message if verbose}
\usage{
.log_msg(verbose, ...)
}
\arguments{
\item{verbose}{Logical, whether to print messages}

\item{...}{Passed to message()}
}
\description{
Print message if verbose
}
\keyword{internal}
