% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_download.R
\name{.log_warn}
\alias{.log_warn}
\title{Print warning if verbose}
\usage{
.log_warn(verbose, ...)
}
\arguments{
\item{verbose}{Logical, whether to print warnings}

\item{...}{Passed to warning()}
}
\description{
Print warning if verbose
}
\keyword{internal}
