% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_download.R
\name{.url_exists}
\alias{.url_exists}
\title{Check if a URL exists}
\usage{
.url_exists(url)
}
\arguments{
\item{url}{Character, URL to check}
}
\value{
Logical, TRUE if URL exists
}
\description{
Check if a URL exists
}
\keyword{internal}
