% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_download.R
\name{download_archive}
\alias{download_archive}
\title{Download and extract a single archive}
\usage{
download_archive(url, destfile = NULL, extract_dir = NULL, verbose = TRUE)
}
\arguments{
\item{url}{\code{character}. URL}

\item{destfile}{\code{character}. Destination file path, or \code{NULL}}

\item{extract_dir}{\code{character} or \code{NULL}. Directory where files will be downloaded and extracted. If \code{NULL}, a temporary directory is used.}

\item{verbose}{\code{logical}. If \code{TRUE}, prints progress messages during download and extraction.}
}
\value{
Path to extraction directory, or NULL if failed
}
\description{
Download and extract a single archive
}
\keyword{internal}
