% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_download.R
\name{download_archives}
\alias{download_archives}
\title{Download and extract multiple archives}
\usage{
download_archives(
  urls,
  destfiles = NULL,
  extract_dir = NULL,
  use_subdirs = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{urls}{\code{character}. Vector of URLs}

\item{destfiles}{\code{character}. Vector of destination file paths, or \code{NULL}}

\item{extract_dir}{\code{character} or \code{NULL}. Directory where files will be downloaded and extracted. If \code{NULL}, a temporary directory is used.}

\item{use_subdirs}{\code{logical}. If \code{TRUE}, create separate subdirectories per archive}

\item{verbose}{\code{logical}. If \code{TRUE}, prints progress messages during download and extraction.}
}
\value{
List of extraction paths
}
\description{
Download and extract multiple archives
}
\keyword{internal}
