% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etalab.R
\name{etalab_construct_url}
\alias{etalab_construct_url}
\title{Construct Etalab data URL}
\usage{
etalab_construct_url(scale = "communes", ...)
}
\arguments{
\item{scale}{Character. Scale of data to retrieve. Must be one of \code{"departements"} or \code{"communes"}.
Default is \code{"communes"}.}

\item{...}{Additional arguments passed to \code{cdg_construct_url()}.}
}
\value{
Character string with the constructed URL.
}
\description{
Constructs a URL to access Etalab data for a specified scale.
}
\examples{
\dontrun{
etalab_construct_url(scale = "departements")
etalab_construct_url(scale = "communes")
}

}
