% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etalab.R
\name{etalab_get_data_url}
\alias{etalab_get_data_url}
\title{Get Etalab data URL for a given INSEE code and data type}
\usage{
etalab_get_data_url(insee_code, data, ...)
}
\arguments{
\item{insee_code}{Character. INSEE code of the geographic unit (commune or department).}

\item{data}{Character. Type of data to retrieve. Must be one of the allowed data types for the given scale.}

\item{...}{Additional arguments passed to \code{etalab_construct_url()} and other internal functions.}
}
\value{
Character string with the constructed URL.
}
\description{
Constructs the download URL for Etalab data based on the INSEE code and requested data type.
}
\details{
The function automatically detects the geographic scale (commune or department) from the INSEE code,
builds the corresponding URL, and verifies the data type is allowed. It constructs the full URL
including the archive prefix and extension.
}
\examples{
\dontrun{
etalab_get_data_url("75056", "batiments")
etalab_get_data_url("69", "voie")
}

}
