% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdg.R
\name{get_data_millesimes}
\alias{get_data_millesimes}
\title{Detect available years (millesimes)}
\usage{
get_data_millesimes(site)
}
\arguments{
\item{site}{\code{character}. The cadastre site to use. Must be one of \code{"pci"} or \code{"etalab"}.}
}
\value{
A \code{character} vector of unique year identifiers (millesimes) found on the site.
}
\description{
Retrieves and returns the list of available "millesimes" (year directories) from a specified data site.
}
\details{
The function queries the base data URL for the specified site and extracts
the list of available millésime directories.
}
\examples{
\dontrun{
years <- list(pci = get_data_millesimes("pci"), etalab = get_data_millesimes("etalab"))
print(years)
}

}
