% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etalab.R
\name{get_etalab}
\alias{get_etalab}
\title{Download cadastre processed datasets from Etalab "bundle" cadastre.data.gouv}
\usage{
get_etalab(id, data = "parcelles", verbose = TRUE)
}
\arguments{
\item{id}{\code{character} or \code{numeric} vector. The INSEE code(s) of the commune(s) or department(s).}

\item{data}{\code{character} vector or \code{list}.
If \code{character}, a vector of layers will be paired with all communes (Cartesian product).
If \code{list}, each element corresponds to a vector of layers for the matching commune.
Must be dataset names returned by \code{get_etalab_layernames("proc")}.}

\item{verbose}{\code{logical}. If \code{TRUE}, prints progress messages.}
}
\value{
An \code{sf} object if a single layer is retrieved, or a named list of \code{sf}
objects if multiple layers are retrieved.
}
\description{
This function downloads one or several dataset layer (e.g., \code{"parcelles"}) for
one or several INSEE identifiers (department or commune) using the Etalab
"bundle" cadastre.data.gouv.
The results are returned as an \code{sf} object. If multiple files are retrieved
(e.g. several communes or layers), they are combined into a single \code{sf}
objects list.
}
\examples{
\dontrun{
# Download parcel geometries for a single commune
parcelles <- get_etalab("72187", data = "parcelles")

# Download several layers for one commune
layers <- get_etalab("72187", data = c("parcelles", "sections"))

# Download parcel geometries for multiple communes
multi <- get_etalab(c("72187", "72032"), data = "parcelles")

# Different layers for each commune
custom <- get_etalab(id = c("72187", "72181"), data = list(c("parcelles", "sections"), "communes"))
}

}
\seealso{
\code{\link[=get_etalab_layernames]{get_etalab_layernames()}}
}
