% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etalab.R
\name{get_etalab_data}
\alias{get_etalab_data}
\title{Download and Read Etalab datasets from server}
\usage{
get_etalab_data(
  commune,
  data,
  millesime = "latest",
  extract_dir = NULL,
  verbose = TRUE
)
}
\arguments{
\item{commune}{\code{character} or \code{numeric} vector. The INSEE code(s) of the commune(s).}

\item{data}{\code{character} vector or \code{list}.
If \code{character}, a vector of layers will be paired with all communes (Cartesian product).
If \code{list}, each element corresponds to a vector of layers for the matching commune.
Must be dataset names returned by \code{\link[=get_etalab_layernames]{get_etalab_layernames()}}.}

\item{millesime}{\code{character}. The version or millesime of the dataset.
Must be on of \code{get_data_millesimes("etalab")}. Default is \code{"latest"}.}

\item{extract_dir}{\code{character} or \code{NULL}. Directory where files will be downloaded and extracted.
If \code{NULL}, a temporary directory is used.}

\item{verbose}{\code{logical}. If \code{TRUE}, prints progress messages.}
}
\value{
An \code{sf} object if a single layer is retrieved, or a named list of \code{sf}
objects if multiple layers are retrieved.
If no valid data is retrieved, the function returns \code{NULL} with a warning.
}
\description{
This function downloads Etalab cadastre data for given communes
and layers, extracts the archives, and reads the GeoJSON files into \code{sf} objects.
}
\examples{
\dontrun{
# Download and read parcels and lieux_dits for one commune
get_etalab_data("72187", c("parcelles", "lieux_dits"))

# Multiple communes with pairwise layers
get_etalab_data(c("72187", "72181"), list(c("parcelles", "lieux_dits"), c("commune")))
}

}
\seealso{
\code{\link[=get_etalab_layernames]{get_etalab_layernames()}}, \code{\link[=get_data_millesimes]{get_data_millesimes()}}
}
