% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pci.R
\name{get_pci_data}
\alias{get_pci_data}
\title{Download and Read PCI datasets from server}
\usage{
get_pci_data(
  x,
  millesime = "latest",
  format = "edigeo",
  extract_dir = NULL,
  verbose = TRUE
)
}
\arguments{
\item{x}{\code{character}. Vector of codes to retrieve URLs for.
Each element must be either a 5-character commune code or a 12-character sheet code.}

\item{millesime}{\code{character}. The version or millesime of the dataset.
Must be on of \code{get_data_millesimes("pci")}. Default is \code{"latest"}.}

\item{format}{\code{character}. The format of the data.
Must be \code{"edigeo"} or \code{"dxf"}. Default is \code{"edigeo"}.}

\item{extract_dir}{\code{character} or \code{NULL}. Directory where files will be downloaded and extracted.
If \code{NULL}, a temporary directory is used.}

\item{verbose}{\code{logical}. If \code{TRUE}, prints progress messages during download and extraction.}
}
\value{
An \code{sf} object or an \code{sf} objects list depending query. Returns \code{NULL} if downloads fail.
}
\description{
This function downloads PCI data for given commune or sheet codes, extracts
the files, and reads them into R as \code{sf} objects (for Edigeo) or another
suitable format (for DXF).
}
\details{
\itemize{
\item Downloads all relevant archives for the provided codes.
\item Extracts files in \code{extract_dir}.
\item Reads Edigeo or DXF files.
}
}
\examples{
\dontrun{
# Download all sheets for a commune
pci_data <- get_pci_data("72187")

# Download a specific sheet
pci_data <- get_pci_data("72181000AB01")

# Multiple communes and sheets
pci_data <- get_pci_data(c("72187", "72181000AB01"))
}

}
\seealso{
\code{\link[=get_data_millesimes]{get_data_millesimes()}}
}
