% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idu.R
\name{idu_build_in_df}
\alias{idu_build_in_df}
\title{Build IDU Codes into a data frame}
\usage{
idu_build_in_df(
  df,
  col_com,
  col_prefix,
  col_section,
  col_number,
  col_dep = NULL
)
}
\arguments{
\item{df}{\code{data.frame} Input data frame.}

\item{col_com}{\code{character}. Name of the commune column.}

\item{col_prefix}{\code{character} Name of the prefix column.}

\item{col_section}{\code{character} Name of the section column.}

\item{col_number}{\code{character} Name of the parcel number column.}

\item{col_dep}{\code{character} or \code{NULL} Name of the department column (optional).}
}
\value{
data.frame. Input data frame with an additional \code{idu} column.
}
\description{
Adds an \code{idu} column to a data frame using parcel components.
}
\details{
Some IDs do not include the department code.
In such cases, it needs to be specified manually,
which is why \code{col_dep} is by defaut set to \code{NULL}.
}
\examples{
df <- data.frame(dep="01", com="002", prefix="0", section="A", number="12")
idu_build_in_df(df, "com", "prefix", "section", "number", "dep")

}
\seealso{
\code{\link[=idu_build]{idu_build()}}
}
