% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idu.R
\name{idu_check}
\alias{idu_check}
\title{Check if a vector contains valid IDUs}
\usage{
idu_check(x)
}
\arguments{
\item{x}{A character vector containing IDU codes to validate.}
}
\value{
Invisibly returns \code{TRUE} if all IDUs are valid; otherwise, stops with an error.
}
\description{
This function checks whether a character vector contains valid INSEE
cadastral parcel identifiers (IDUs). It can either return a logical
vector indicating validity or raise an error if invalid entries are found.
}
\details{
A valid IDU (INSEE cadastral identifier) must satisfy all of the following:
\itemize{
\item Be a non-missing, non-empty string of length 14.
\item The first two characters: department code — digits (0–9) or letters 'A'/'B'.
\item The next three characters: commune code — digits (0–9).
\item The next three characters: prefix — digits (0–9).
\item The next two characters: section — digits (0–9) or uppercase letters (A–Z).
\item The last four characters: parcel number — digits (0–9).
\item Contain no lowercase letters or special characters.
}
}
\examples{
\dontrun{
# Valid IDU
idu_check("01001000AA0123")

# Multiple values
idu_check(c("01001000AA0123", "AB12300000A0123"))

# Invalid example (will throw an error)
idu_check(c("01001000AA0123", "12345"))
}

}
