% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idu.R
\name{idu_fill_df}
\alias{idu_fill_df}
\title{Fill a Data Frame with IDU Codes}
\usage{
idu_fill_df(df, col_com, col_prefix, col_section, col_number, col_dep = NULL)
}
\arguments{
\item{df}{\code{data.frame} Input data frame.}

\item{col_com}{\code{character}. Name of the commune column.}

\item{col_prefix}{\code{character} Name of the prefix column.}

\item{col_section}{\code{character} Name of the section column.}

\item{col_number}{\code{character} Name of the parcel number column.}

\item{col_dep}{\code{character} or \code{NULL} Name of the department column (optional).}
}
\value{
data.frame. Input data frame with an additional \code{idu} column.
}
\description{
Adds an \code{idu} column to a data frame using parcel components.
}
\examples{
df <- data.frame(dep="01", com="002", prefix="0", section="A", number="12")
idu_fill_df(df, "com", "prefix", "section", "number", "dep")

}
\seealso{
\code{\link[=idu_build]{idu_build()}}
}
