% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idu_from_pdf.R
\name{idu_from_pdf}
\alias{idu_from_pdf}
\title{Extract IDUs from cadastral matrices in PDF format}
\usage{
idu_from_pdf(path = NULL, interactive = FALSE)
}
\arguments{
\item{path}{\code{character}; vector of file paths to cadastral matrix PDF files.}

\item{interactive}{\code{boolean}; if \code{TRUE}, prompts the user to select PDF
files interactively.}
}
\value{
A character vector of all extracted IDU codes.
}
\description{
When cadastral matrices are requested from the DGFIP, they are provided
as PDF files.
This function parses one or several PDFs to extract all IDU codes.
}
\examples{
\dontrun{

idu_from_pdf(c("C://cad1.pdf", "C://cad2.pdf"))

}
}
