% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idu.R
\name{idu_get_feuille}
\alias{idu_get_feuille}
\title{Get feuille IDs from IDU codes}
\usage{
idu_get_feuille(idu, group_by_insee = FALSE)
}
\arguments{
\item{idu}{A \code{character} vector of valid IDU codes.}
}
\value{
A \code{character} vector of feuille IDs corresponding to the input IDUs.
}
\description{
This function takes one or more IDU codes, validates them, extracts their
INSEE and feuille components, downloads feuille data from Etalab, and returns
the matching feuille IDs.
}
\examples{
\dontrun{
# Get feuille IDs for a set of IDU codes
idu_get_feuille(c("12345ABCDE6789", "54321ZZZZZ1234"))
}

}
