% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idu.R
\name{idu_get_feuille_in_df}
\alias{idu_get_feuille_in_df}
\title{This function detects the column containing IDU codes in a given
\code{data.frame} or \code{sf} object, validates the codes, and retrieves the
corresponding feuille identifiers from the Etalab reference dataset.}
\usage{
idu_get_feuille_in_df(df)
}
\arguments{
\item{df}{A \code{data.frame} or \code{sf} object containing at least one column
with valid IDU codes.}
}
\value{
The input \code{df} with an additional column \code{feuille_id}
containing the feuille identifiers corresponding to each IDU.
}
\description{
This function detects the column containing IDU codes in a given
\code{data.frame} or \code{sf} object, validates the codes, and retrieves the
corresponding feuille identifiers from the Etalab reference dataset.
}
\examples{
\dontrun{
df <- data.frame(
  idu_code = c("12345ABCDE6789", "54321ZZZZZ1234")
)
idu_get_feuille_in_df(df)
}

}
\seealso{
\code{\link[=idu_get_feuille]{idu_get_feuille()}}, \code{\link[=idu_detect_in_df]{idu_detect_in_df()}}
}
