% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idu.R
\name{idu_get_lieudit}
\alias{idu_get_lieudit}
\title{Retrieve "lieu-dit" names for given IDUs}
\usage{
idu_get_lieudit(idu)
}
\arguments{
\item{idu}{A \code{character} vector of valid IDU codes.}
}
\value{
A \code{data.frame} containing the split IDU parts and an additional
\code{lieudit} column with the matched lieu-dit names.
}
\description{
This function takes one or more valid IDU codes, retrieves parcel and lieu-dit
data from Etalab, performs a spatial join to determine the lieu-dit for each IDU,
and returns a data.frame mapping IDUs to their lieu-dit names.
}
\examples{
\dontrun{
idu_get_lieudit(c("721870000A0001", "721870000A0002"))
}

}
\keyword{internal}
