% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idu.R
\name{idu_get_name_in_df}
\alias{idu_get_name_in_df}
\title{Append location names to a data.frame containing IDU codes}
\usage{
idu_get_name_in_df(df, loc = c("both", "reg", "dep", "com"), cog_field = "NCC")
}
\arguments{
\item{df}{A \code{data.frame} containing an IDU column.}

\item{loc}{A \code{character} string specifying which location levels to include.
One of \code{"both"}, \code{"reg"}, \code{"dep"}, or \code{"com"}.}

\item{cog_field}{\code{character}. Défaut is \code{"NCC"}.
The name of the field in the reference datasets to use.}
}
\value{
A \code{data.frame} with new location name columns appended.
}
\description{
This function detects the column containing IDU codes in a data.frame,
then uses \code{\link{idu_get_name}} to append the corresponding
region, department, and/or commune names.
}
\examples{
# my_df <- data.frame(idu = c("12345ABCDE6789", "54321ZZZZZ1234"))
# idu_get_name_in_df(my_df, loc = "both")

}
