% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idu.R
\name{idu_get_parcelle}
\alias{idu_get_parcelle}
\title{Retrieve Parcelles from IDU Codes}
\usage{
idu_get_parcelle(idu, filter = TRUE)
}
\arguments{
\item{idu}{\code{character}. Vector of IDU codes.}

\item{filter}{\code{logical}. Defaut is \code{TRUE}.
If \code{TRUE}, only returns the parcels matching the IDU codes.}
}
\value{
A \code{sf} object. Filtered parcel geometries.
}
\description{
Downloads and filters parcel data corresponding to a vector of IDU codes.
}
\seealso{
\code{\link[=idu_split(), get_quick_etalab(), lieudits_for_parcelles]{idu_split(), get_quick_etalab(), lieudits_for_parcelles()}}
}
