% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idu.R
\name{idu_split}
\alias{idu_split}
\title{Split IDU into Its Components}
\usage{
idu_split(
  idu,
  col_dep = "dep",
  col_com = "com",
  col_prefix = "prefix",
  col_section = "section",
  col_number = "number"
)
}
\arguments{
\item{idu}{\code{character}
A vector of IDU codes.}

\item{col_dep}{\code{character}
Name for the department column (default \code{"dep"}).}

\item{col_com}{\code{character}
Name for the commune column (default \code{"com"}).}

\item{col_prefix}{\code{character}
Name for the prefix column (default \code{"prefix"}).}

\item{col_section}{\code{character}
Name for the section column (default \code{"section"}).}

\item{col_number}{\code{character}
Name for the number column (default \code{"number"}).}
}
\value{
data.frame. Columns corresponding to the components of the IDU.
}
\description{
Splits a 14-character parcel IDU into department, commune, prefix, section, and number.
}
\examples{
idu_split("0100200A0012")

}
