% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idu.R
\name{lieudits_for_parcelles}
\alias{lieudits_for_parcelles}
\title{Associate Lieudits to Parcelles}
\usage{
lieudits_for_parcelles(parcelles, lieudits, col_idu = "id")
}
\arguments{
\item{parcelles}{\code{sf} object.
Parcel geometries with \code{id} column.}

\item{lieudits}{\code{sf} object.
Lieudit geometries with \code{id} and \code{nom} columns.}
}
\value{
sf object. Parcelles with an additional \code{lieudit} column.
}
\description{
Computes centroid for each parcel, intersects with lieudits, and merges the lieudit name into the parcels table.
}
\seealso{
\code{\link[=idu_get_parcelle]{idu_get_parcelle()}}
}
