% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pci.R
\name{pci_detect_feuilles}
\alias{pci_detect_feuilles}
\title{Détecter les feuilles PCI disponibles pour une commune}
\usage{
pci_detect_feuilles(commune, format = "edigeo", skip = FALSE, ...)
}
\arguments{
\item{commune}{\code{character}.
Code INSEE de la commune (ex: "02120").}

\item{...}{Arguments supplémentaires passés à \code{\link{pci_construct_url}}.}
}
\value{
\code{character}.
Un vecteur des noms des feuilles PCI disponibles.
}
\description{
Cette fonction retourne la liste des feuilles disponibles pour une commune donnée dans le PCI.
}
\details{
La fonction construit l'URL correspondant à la commune
et utilise \code{cdg_detect_links()} pour récupérer les feuilles disponibles.
}
\seealso{
\code{\link[=pci_construct_url(), cdg_construct_commune(), cdg_detect_links]{pci_construct_url(), cdg_construct_commune(), cdg_detect_links()}}
}
