% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pci.R
\name{pci_get_dep_url}
\alias{pci_get_dep_url}
\title{Construire l'URL d'un département PCI}
\usage{
pci_get_dep_url(departement, ...)
}
\arguments{
\item{departement}{\code{character}.
Code du département à utiliser. Doit appartenir à la liste \code{Rsequoia2::departements_2024$DEP}.}

\item{...}{Arguments supplémentaires passés à \code{\link{pci_construct_url}}.}

\item{zip_prefix}{\code{character}.
Préfixe du fichier zip. Par défaut \code{"dep"}.}

\item{zip_ext}{\code{character}.
Extension du fichier zip. Par défaut \code{".zip"}.}
}
\value{
\code{character}.
Une URL construite sous forme de chaîne de caractères.
}
\description{
Cette fonction génère l'URL d'accès aux données PCI pour un département donné.
}
\details{
La fonction vérifie que le département est valide avant de construire l'URL.
}
\seealso{
\code{\link[=pci_construct_url(), cdg_aggr_url]{pci_construct_url(), cdg_aggr_url()}}
}
