% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdg.R
\name{cdg_download_archive}
\alias{cdg_download_archive}
\title{Download and optionally extract an archive from a URL}
\usage{
cdg_download_archive(
  url,
  destfile = NULL,
  extract = FALSE,
  extract_dir = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{url}{\code{character}. URL of the archive to download.}

\item{destfile}{\code{character}.
Optional path to save the downloaded file.
If \code{NULL}, a temporary file with an appropriate extension is created.}

\item{extract}{\code{logical}. Defaults is \code{FALSE}.
Whether to extract the archive after download.}

\item{extract_dir}{\code{character}.
Directory where the archive will be extracted.
If \code{extract} is \code{TRUE}. If \code{NULL}, a temporary directory is created.}

\item{overwrite}{\code{logical}. Defaults is \code{FALSE}.
Whether to overwrite the destination file if it already exists.}
}
\value{
If \code{extract} is \code{TRUE}, returns the path to the extraction directory.
Otherwise, returns a \code{list} with components:
\item{file}{Path to the downloaded file.}
\item{contents}{Metadata about the archive contents (as returned by \code{archive::archive}).}
}
\description{
Downloads a file from the specified \code{url} to a destination file. Optionally,
the downloaded archive can be extracted to a given directory.
}
\seealso{
\code{\link[archive:archive_extract]{archive::archive_extract()}}, \code{\link[archive:archive]{archive::archive()}}
}
