% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etalab.R
\name{etalab_construct_url}
\alias{etalab_construct_url}
\title{Construct Etalab data URL}
\usage{
etalab_construct_url(scale = "communes", ...)
}
\arguments{
\item{scale}{\code{character}. Default is \code{"communes"}.
Scale of data to retrieve. Must be one of \code{"departements"} or \code{"communes"}.}

\item{...}{Additional arguments passed to \code{\link{cdg_construct_url}}.}
}
\value{
A \code{character} string with the constructed URL.
}
\description{
Constructs a URL to access Etalab data for a specified scale.
}
\examples{
\dontrun{
etalab_construct_url(scale = "departements")
etalab_construct_url(scale = "communes")
}

}
\seealso{
\code{\link[=cdg_construct_url]{cdg_construct_url()}}
}
