% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quick_etalab.R
\name{get_quick_etalab}
\alias{get_quick_etalab}
\title{Download and return a specific cadastre layer from Etalab}
\usage{
get_quick_etalab(insee_code, data = "parcelles", ...)
}
\arguments{
\item{insee_code}{\code{character} or \code{numeric}.
Vector of INSEE codes for the communes to download.}

\item{data}{\code{character}.
Name of the data layer to retrieve. Default is \code{"parcelles"}.
See available data with \code{\link{cfg_get_data}}}

\item{...}{Additional arguments passed to \code{get_cadastre_etalab()}.}
}
\value{
\code{sf} object. The requested cadastral layer with unique features.
}
\description{
Retrieves cadastral data for one or more INSEE codes from Etalab,
optionally specifying the type of data layer (e.g., "parcelles" or "lieudit").
}
\seealso{
\code{\link[=get_cadastre_etalab(), cfg_get_data]{get_cadastre_etalab(), cfg_get_data()}}
}
