% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idu.R
\name{idu_build}
\alias{idu_build}
\title{Build a Unique ID (IDU) for a Parcel}
\usage{
idu_build(dep = NULL, com, prefix = "0", section, numero)
}
\arguments{
\item{dep}{\code{character} or \code{NULL} Department code (optional).}

\item{com}{\code{character} Commune code.}

\item{prefix}{\code{character} Prefix code (default \code{"0"}).}

\item{section}{\code{character} Section code.}

\item{numero}{\code{character} Parcel number.}
}
\value{
\code{character} A 14-character unique ID (IDU).
}
\description{
Constructs a standardized 14-character ID for a parcel based on department, commune, prefix, section, and number.
}
\details{
Some IDs do not include the department code.
In such cases, it needs to be specified manually,
which is why \code{dep} is by defaut set to \code{NULL}
}
\examples{
idu_build(dep = "01", com = "002", prefix = "0", section = "A", numero = "12")

}
