% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idu.R
\name{idu_check}
\alias{idu_check}
\title{Check if a vector contains valid IDU codes}
\usage{
idu_check(x)
}
\arguments{
\item{x}{A character vector to validate.}
}
\value{
A logical vector, TRUE for valid IDU entries.
}
\description{
Check if a vector contains valid IDU codes
}
\details{
An IDU is defined as:
\itemize{
\item A \code{character} vector of length 14 for all entries
\item First 5 characters are digits (0–9)
\item Next 5 characters are uppercase letters (A–Z) or digits (0–9)
\item Last 4 characters are digits (0–9)
\item No missing values (NA) or empty strings
\item No lowercase letters or special characters
}
}
