% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idu.R
\name{idu_check}
\alias{idu_check}
\title{Check if a vector contains valid IDU codes}
\usage{
idu_check(x)
}
\arguments{
\item{x}{A character vector to validate.}
}
\value{
A logical vector, TRUE for valid IDU entries.
}
\description{
Check if a vector contains valid IDU codes
}
\details{
An IDU is defined as:
\itemize{
\item A \code{character} vector of length 14 for all entries
\item First 1 characters: department -> digits (0–9)
\item Next 1 characters: department -> digits (0–9) or uppercase letters 'A'/'B'
\item Next 3 characters: commune -> digits (0–9)
\item Next 3 characters: prefixe -> digits (0–9)
\item Next 2 characters: section -> digits (0–9) or uppercase letters (A–Z)
\item Last 4 characters: numero -> digits (0–9)
\item No missing values (NA) or empty strings
\item No lowercase letters or special characters
}
}
