% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pci.R
\name{pci_construct_url}
\alias{pci_construct_url}
\title{Construire une URL pour les données PCI}
\usage{
pci_construct_url(format = "edigeo", scale = "feuilles", ...)
}
\arguments{
\item{format}{\code{character}.
Format des données, parmi \code{"dxf"} ou \code{"edigeo"}.
Par défaut \code{"edigeo"}.}

\item{scale}{\code{character}.
Échelle des données, parmi \code{"departements"} ou \code{"feuilles"}.
Par défaut \code{"feuilles"}.}

\item{...}{Arguments supplémentaires passés à \code{cdg_construct_url()}.}
}
\value{
\code{character}.
Une URL construite sous forme de chaîne de caractères.
}
\description{
Cette fonction construit une URL complète pour accéder aux données PCI
en fonction du format et de l'échelle choisis.
}
\seealso{
\code{\link[=cdg_construct_url]{cdg_construct_url()}}
}
