% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pci.R
\name{pci_detect_feuilles}
\alias{pci_detect_feuilles}
\title{Detect available PCI sheets for a municipality}
\usage{
pci_detect_feuilles(commune, format = "edigeo", skip = FALSE, ...)
}
\arguments{
\item{commune}{\code{character}.
INSEE code of the municipality (e.g., "02120").}

\item{format}{\code{character}.
PCI format (default: \code{"edigeo"}).}

\item{skip}{\code{logical}.
If \code{TRUE}, removes the \verb{edigeo-} or \verb{dxf-} prefix and the \code{.tar.bz2} suffix from the filenames.}

\item{...}{Additional arguments passed to \code{\link{pci_construct_url}}.}
}
\value{
\code{character}.
A vector of the PCI sheet names available.
}
\description{
This function returns the list of available PCI sheets for a given municipality.
}
\details{
The function builds the URL for the municipality and uses \code{cdg_detect_links()}
to retrieve the list of available sheets.
}
\seealso{
\code{\link[=pci_construct_url]{pci_construct_url()}}, \code{\link[=cdg_construct_commune]{cdg_construct_commune()}}, \code{\link[=cdg_detect_links]{cdg_detect_links()}}
}
