% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pci.R
\name{pci_get_feuille_url}
\alias{pci_get_feuille_url}
\title{Obtenir l'URL d'une feuille PCI pour une commune donnée}
\usage{
pci_get_feuille_url(
  commune,
  feuille = NULL,
  format = "edigeo",
  skip = TRUE,
  ...
)
}
\arguments{
\item{commune}{\code{character}.
Code INSEE de la commune (ex: "02120").}

\item{feuille}{\code{character}.
Nom de la feuille PCI (sans préfixe ni extension).}

\item{...}{Arguments supplémentaires passés à \code{pci_construct_url()}.}

\item{bz2_prefix}{\code{character}.
Préfixe du fichier archive (par défaut \code{"edigeo-"}).}

\item{bz2_ext}{\code{character}.
Extension du fichier archive (par défaut \code{".tar.bz2"}).}
}
\value{
\code{character}.
Une URL complète vers la feuille PCI.
}
\description{
Cette fonction construit l'URL d'accès à une feuille PCI spécifique pour une commune.
}
\details{
La fonction vérifie que la feuille demandée existe parmi celles détectées pour la commune,
sinon elle génère une erreur.
}
\seealso{
\code{\link[=pci_construct_url(), pci_detect_feuilles(), cdg_construct_commune(), cdg_aggr_url]{pci_construct_url(), pci_detect_feuilles(), cdg_construct_commune(), cdg_aggr_url()}}
}
