% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_geojson}
\alias{read_geojson}
\title{Read GeoJSON file(s) as an sf object}
\usage{
read_geojson(path)
}
\arguments{
\item{path}{\code{character} Path to a GeoJSON/JSON file or a directory containing such files.}
}
\value{
An \code{sf} object representing the spatial features read from the GeoJSON file(s),
or a list of \code{sf} objects if aggregation fails.
}
\description{
This function reads a single GeoJSON/JSON file or all GeoJSON/JSON files
in a directory and attempts to aggregate them into a single \code{sf} object.
}
\seealso{
\code{\link[sf]{st_read}} for reading spatial vector data.
}
